+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Warn that using Rider as a target IntelliJ Platform with `useInstaller = true` is currently not supported, please set `useInstaller = false` instead. JetBrains/intellij-platform-gradle-plugin#1852
 - Add `ide(type: Provider<*>, version: Provider<String>, useInstaller: Provider<Boolean>)` overload to the `pluginVerification.ides` block.
 - Warn that since the IntelliJ Platform version `2025.1` (build `251`), the required Kotlin version is `2.0.0` or higher.
 - Load the `com.intellij` bundled module by default for all IntelliJ Platform types